refactor: simplify AST object option conversions
