+++ crates/rspack_plugin_javascript/src/utils/mod.rs
@@ -1,4 +1,3 @@
-pub mod ast_object;
 pub mod eval;
 pub mod mangle_exports;
 pub mod object_properties;
